+++ testing/execution.ts
@@ -456,7 +456,10 @@ async function executeFormulaOrSyncWithRawParamsInVM<T extends FormulaSpecificat
       break;
     }
     case FormulaType.Metadata: {
-      params = parseMetadataFormulaParams(rawParams) as ParamValues<ParamDefs>;
+      // Interstingly we don't need special handling for the formula context dict (the optional second arg
+      // to an autocomplete metadata formula), because at execution time it gets passed as a serialized
+      // JSON string anyway which is already parsed by the compiled pack definition.
+      params = rawParams as ParamValues<ParamDefs>;
       break;
     }
     case FormulaType.SyncUpdate: {
@@ -504,7 +507,10 @@ export async function executeFormulaOrSyncWithRawParams<T extends FormulaSpecifi
       break;
     }
     case FormulaType.Metadata: {
-      params = parseMetadataFormulaParams(rawParams) as ParamValues<ParamDefs>;
+      // Interstingly we don't need special handling for the formula context dict (the optional second arg
+      // to an autocomplete metadata formula), because at execution time it gets passed as a serialized
+      // JSON string anyway which is already parsed by the compiled pack definition.
+      params = rawParams as ParamValues<ParamDefs>;
       break;
     }
     case FormulaType.SyncUpdate: {
@@ -515,11 +521,6 @@ export async function executeFormulaOrSyncWithRawParams<T extends FormulaSpecifi
       ensureUnreachable(formulaSpecification);
   }
   return findAndExecutePackFunction(params, formulaSpecification, manifest, executionContext, syncUpdates);
-}
-
-function parseMetadataFormulaParams(rawParams: string[]): string[] {
-  const [search = '', formulaContext = '{}'] = rawParams;
-  return [search, JSON.parse(formulaContext)];
 }
 
 /**